refactor: restructure project for python
