+++ Dockerfile
@@ -2,4 +2,4 @@ FROM python:3.11-slim
 WORKDIR /app
 COPY . .
 RUN pip install --no-cache-dir boto3
-CMD ["python", "main.py"]
+CMD ["python", "-m", "infra_cleanup.main"]
