+++ tests/__init__.py
@@ -1,1 +1,3 @@
-
+import sys
+from pathlib import Path
+sys.path.append(str(Path(__file__).resolve().parents[1] / "src"))
